Renames DebugProvider to DebugInternal and and adds DebugProvider class
