+++ src/System.Private.CoreLib/shared/System/Diagnostics/Debug.cs
@@ -15,14 +15,31 @@ namespace System.Diagnostics
     {
         private static readonly object s_lock = new object();
 
-        private static DebugProvider s_provider = new DebugProvider();
-
-        public static DebugProvider SetProvider(DebugProvider provider)
+        private static DebugInternal s_internal = new DebugInternal();
+        // Enables Debug to get wired with Trace Listeners
+        public static void ReplaceInternal(DebugInternal internal)
         {
-            if (provider == null)
-                throw new ArgumentNullException(nameof(provider));
+            if (internal == null)
+                throw new ArgumentNullException(nameof(internal));
 
-            return Interlocked.Exchange(ref s_provider, provider);
+            Interlocked.Exchange(ref s_internal, internal);
+        }
+
+        private static DebugProvider s_provider = new DebugProvider();
+        // Enables (WPF) user with option to override ShowDialog and Write logic
+        public static DebugProvider Provider
+        {
+            get
+            {
+                return s_provider;
+            }
+            set
+            {
+                if (value == null)
+                    throw new ArgumentNullException(nameof(value));
+
+                Interlocked.Exchange(ref s_provider, value);
+            }
         }
 
         public static bool AutoFlush { get { return true; } set { } }
@@ -40,7 +57,7 @@ namespace System.Diagnostics
                 lock (s_lock)
                 {
                     t_indentLevel = value < 0 ? 0 : value;
-                    s_provider.OnIndentLevelChanged(t_indentLevel);
+                    s_internal.OnIndentLevelChanged(t_indentLevel);
                 }
             }
         }
@@ -57,7 +74,7 @@ namespace System.Diagnostics
                 lock (s_lock)
                 {
                     s_indentSize = value < 0 ? 0 : value;
-                    s_provider.OnIndentSizeChanged(s_indentSize);
+                    s_internal.OnIndentSizeChanged(s_indentSize);
                 }
             }
         }
@@ -178,13 +195,13 @@ namespace System.Diagnostics
         [System.Diagnostics.Conditional("DEBUG")]
         public static void WriteLine(string message)
         {
-            s_provider.WriteLine(message);
+            s_internal.WriteLine(message);
         }
 
         [System.Diagnostics.Conditional("DEBUG")]
         public static void Write(string message)
         {
-            s_provider.Write(message);
+            s_internal.Write(message);
         }
 
         [System.Diagnostics.Conditional("DEBUG")]
