+++ src/System.Private.CoreLib/shared/System/Diagnostics/DebugProvider.Unix.cs
@@ -38,7 +38,7 @@ namespace System.Diagnostics
             }
         }
 
-        private static void WriteCore(string message)
+        public virtual void Write(string message)
         {
             WriteToDebugger(message);
