+++ src/System.Private.CoreLib/shared/System/Diagnostics/DebugProvider.cs
@@ -8,43 +8,10 @@
 namespace System.Diagnostics
 {
     /// <summary>
-    /// Provides default implementation for Write and ShowDialog methods in Debug class.
+    /// Enables Debug class with option to customize ShowDialog and Write
     /// </summary>
     public partial class DebugProvider
     {
-        public virtual void Write(string message)
-        {
-            lock (s_lock)
-            {
-                if (message == null)
-                {
-                    s_WriteCore(string.Empty);
-                    return;
-                }
-                if (_needIndent)
-                {
-                    message = GetIndentString() + message;
-                    _needIndent = false;
-                }
-                s_WriteCore(message);
-                if (message.EndsWith(Environment.NewLine))
-                {
-                    _needIndent = true;
-                }
-            }
-        }
-        
-        public virtual void WriteLine(string message)
-        {
-            Write(message + Environment.NewLine);
-        }
-
-        public virtual void OnIndentLevelChanged(int indentLevel) { }
-
-        public virtual void OnIndentSizeChanged(int indentSize) { }
-
-        private static readonly object s_lock = new object();
-
         private sealed class DebugAssertException : Exception
         {
             internal DebugAssertException(string stackTrace) :
@@ -62,22 +29,5 @@ namespace System.Diagnostics
             {
             }
         }
-
-        private bool _needIndent = true;
-
-        private string _indentString;
-
-        private string GetIndentString()
-        {
-            int indentCount = Debug.IndentSize * Debug.IndentLevel;
-            if (_indentString?.Length == indentCount)
-            {
-                return _indentString;
-            }
-            return _indentString = new string(' ', indentCount);
-        }
-
-        // internal and not readonly so that the tests can swap this out.
-        internal static Action<string> s_WriteCore = WriteCore;
     }
 }
