+++ src/System.Private.CoreLib/src/System/Diagnostics/DebugProvider.Windows.cs
@@ -33,10 +33,10 @@ namespace System.Diagnostics
                 Environment.FailFast(ex.Message, ex, errorSource);
             }
         }
-
+    
         private static readonly object s_ForLock = new object();
 
-        private static void WriteCore(string message)
+        public virtual void Write(string message)
         {
             // really huge messages mess up both VS and dbmon, so we chop it up into 
             // reasonable chunks if it's too big. This is the number of characters
